requirements.txt
affine==2.4.0
aiobotocore==2.21.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
alabaster==1.0.0
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bioio==1.1.0
bioio-base==1.0.6
bioio-ome-tiff==1.0.2
bleach==6.2.0
blinker==1.9.0
bokeh==3.5.2
botocore==1.37.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dask==2025.3.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dicomweb-client==0.59.3
distributed==2025.3.0
docutils==0.21.2
einops==0.8.1
elementpath==4.8.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
fastremap==1.15.1
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
geojson==3.2.0
Glymur==0.13.6
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imagecodecs==2024.9.22
imageio==2.27.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib-metadata==4.13.0
importlib-resources==5.13.0
instanseg==0.0.2
instanseg-torch==0.0.8
ipykernel==6.29.5
ipython==8.36.0
isoduration==20.11.0
itkwasm==1.0b189
itkwasm-downsample==1.6.0
itkwasm-downsample-wasi==1.6.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.43.0
locket==1.0.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
MedPy==0.5.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ngff-zarr==0.13.0
notebook_shim==0.2.4
nptyping==1.4.4
numba==0.60.0
numcodecs==0.13.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ome-types==0.6.0
ome-zarr==0.10.3
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
openslide-python==1.4.1
overrides==7.7.0
packaging==24.2
palettable==3.3.3
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic_core==2.33.1
pydicom==3.0.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-argparse==1.7.0
rpds-py==0.24.0
s3fs==2025.3.2
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
shapely==2.0.6
SimpleITK==2.4.0
simsimd==6.2.1
six==1.17.0
slideio==2.7.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
stringzilla==3.12.3
sympy==1.13.1
tblib==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiatoolbox==1.6.0
tifffile==2024.9.20
tiffslide==2.5.0
timm==1.0.12
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
torch==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
ttach==0.0.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
typish==1.9.3
tzdata==2025.2
umap-learn==0.5.7
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==2.4.0
wasmtime==31.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsidicom==0.22.0
xarray==2025.3.1
xmlschema==4.0.1
xsdata==24.3.1
xyzservices==2025.4.0
yarl==1.19.0
zarr==2.17.2
zict==3.0.0
zipp==3.21.0
